configuring tests
