+++ client/app.jsx
@@ -1,52 +1,6 @@
 import React from 'react';
-import Overall from './components/overall.jsx';
-import Reviews from './components/reviews.jsx';
+import App from './render.jsx';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Route, Link } from "react-router-dom";
-
-
-class App extends React.Component{
-  constructor(props) {
-    super(props);
-    this.state = {
-      reviews: [],
-      items: [],
-    }
-    this.getReviews = this.getReviews.bind(this);
-    this.getItems = this.getItems.bind(this);
-  }
-
-  componentDidMount() {
-    this.getReviews();
-    this.getItems();
-    console.log(window.location.pathname);
-  }
-
-  getReviews() {
-    fetch('/allReviews') //need to send window.location.pathway as url but how does it receive on server side?
-    .then(data => data.json())
-    .then(reviews => this.setState({reviews: reviews}))
-  }
-
-  getItems() {
-    fetch('/allItems')
-    .then(data => data.json())
-    .then(items => this.setState({items: items}))
-  }
-
-
-
-
-
-  render() {
-    return (
-      <div>
-        <Overall items={this.state.items} reviews={this.state.reviews}/>
-        {this.state.reviews.map(review => <Reviews header={review.header} date={review.date} userFirstName={review.userFirstName} userLastInitial={review.userLastInitial} review={review.review} stars={review.stars} tips={review.tips}/>)}
-      </div>
-    )
-  }
-
-}
 
 ReactDOM.render(<App />, document.getElementById('app'));
+
