+++ package.json
@@ -1,5 +1,16 @@
 {
   "name": "",
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "jsx"
+    ],
+    "moduleDirectories": [
+      "node_modules",
+      "bower_components",
+      "shared"
+    ]
+  },
   "version": "",
   "description": "",
   "author": "",
@@ -14,8 +25,11 @@
     "test": "jest"
   },
   "dependencies": {
+    "D": "^1.0.0",
     "body-parser": "^1.18.3",
+    "chai": "^4.2.0",
     "express": "^4.16.4",
+    "mocha": "^6.0.2",
     "moment": "^2.24.0",
     "mysql": "^2.16.0",
     "react": "^16.8.5",
@@ -23,13 +37,17 @@
     "react-ratings-declarative": "^3.4.1",
     "react-router-dom": "^5.0.0",
     "react-star-ratings": "^2.3.0",
+    "react-test-renderer": "^16.8.6",
+    "sinon": "^7.3.1",
     "styled-components": "^4.2.0",
     "webpack": "^4.29.6"
   },
   "devDependencies": {
     "@babel/core": "^7.4.0",
+    "@babel/polyfill": "^7.4.0",
     "@babel/preset-env": "^7.4.2",
     "@babel/preset-react": "^7.0.0",
+    "babel-jest": "^24.5.0",
     "babel-loader": "^8.0.5",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.11.2",
@@ -37,6 +55,9 @@
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "eslint-plugin-react": "^7.12.4",
     "jest": "^24.5.0",
+    "jest-cli": "^24.5.0",
+    "postcss-import": "^8.0.2",
+    "puppeteer": "^1.14.0",
     "webpack-cli": "^3.3.0"
   }
 }
