+++ test/form.spec.js
@@ -1,0 +1,20 @@
+
+import React from 'react';
+import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
+import { configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+import sinon from 'sinon';
+configure({ adapter: new Adapter() });
+import Overall from '../client/components/overall.jsx';
+import App from '../client/render.jsx'
+
+
+
+describe('<App />', () => {
+  it('calls componentDidMount', () => {
+    sinon.spy(App.prototype, 'componentDidMount');
+    const wrapper = mount(<App />);
+    expect(App.prototype.componentDidMount).to.have.property('callCount', 1);
+  });
+})
